Clean code improvements

+++ GolfImprovementMeasurement/Models/GolfRound.cs
@@ -3,5 +3,5 @@ namespace GolfImprovementMeasurement.Models;
 internal sealed record GolfRound(
     int DaysSinceReference,
     int NumberOfShots,
-    decimal CourseCondition,
-    decimal CourseMultiplier);
+    double CourseCondition,
+    double CourseMultiplier);
